Separate `LevelBuilder` backend from domain types
