+++ src/game/level/builder/mod.rs
@@ -1,4 +1,3 @@
-pub mod backend;
 pub mod error;
 /// Contains declarations and general helper structs for the [`LevelBuilder`]
 /// Implementations of [`LevelBuilder`] are in layout and builder submodules.
